requirements.txt
fpdf2>=2.8.5
pillow>=12.0.0
pydrive2>=1.21.3
